Kladd redesign migrer arbeidsliste til huskelapp
